feat(exec): a SET sorts by its member bitmask, matching MySQL
Found by the sakila real-data differential (#261): Pintail ordered
film.special_features alphabetically where MySQL orders a SET by its
member bitmask. Same architecture as the ENUM ordinal work:

- Value::Enum's index widens to u64 so it carries a SET's mask (up to
  64 members) as well as an ENUM's declaration index; derived Ord is
  unchanged.
- The probe attaches declared SET members beside ENUM labels; the
  schema column, scan spec, decode paths (arena, dictionary template,
  row-shaped, prewhere, adopt_chunk) and the two-pass group-key rebuild
  all thread them in parallel with enum labels.
- StrColumn::declared_ordinal and the free ordinal_value helper derive
  either kind's ordinal; an undeclared member refuses and the value
  stays text, exactly like an undeclared ENUM label.
- MIN/MAX and comparisons keep STRING semantics - measured on MySQL
  8.4, which returns 'Trailers,Commentaries' (lexically last) as MAX,
  not the highest mask - so only sorting changed.

Verified byte-exact against live sakila on five SET shapes (grouped
tie-breaks included) and pinned by ten in-process variations in
crates/pintail-exec/tests/set_ordinal.rs. Oracle: 1003 cases green.

+++ crates/pintail-types/src/value.rs
@@ -185,8 +185,9 @@ pub enum Value {
     /// Derived `Ord` compares `index` before `label`, which is the ordering
     /// `MySQL` uses.
     Enum {
-        /// One-based declaration index, matching `MySQL`'s ordinal.
-        index: u16,
+        /// One-based declaration index for an ENUM, or the member bitmask
+        /// for a SET - both are what `MySQL` sorts the type by.
+        index: u64,
         /// Declared label, and what the value displays as.
         label: String,
     },
@@ -240,7 +241,7 @@ impl Value {
 
     /// The declaration index when this is an ENUM.
     #[must_use]
-    pub const fn enum_index(&self) -> Option<u16> {
+    pub const fn enum_index(&self) -> Option<u64> {
         match self {
             Self::Enum { index, .. } => Some(*index),
             _ => None,
